--- service-user/src/main/java/com/bank/serviceuser/ServiceUserApplication.java
@@ -1,13 +1,0 @@
-package com.bank.serviceuser;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ServiceUserApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ServiceUserApplication.class, args);
-	}
-
-}
