--- service-user/src/main/java/com/bank/serviceuser/config/ModelMapperConfig.java
@@ -1,14 +1,0 @@
-package com.bank.serviceuser.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-}
