--- service-user/src/main/java/com/bank/serviceuser/dto/UserResponseDTO.java
@@ -1,11 +1,0 @@
-package com.bank.serviceuser.dto;
-
-import lombok.Data;
-
-@Data
-public class UserResponseDTO {
-    private Long id;
-    private String nombreCompleto;
-    private String email;
-    private String telefono;
-}
