--- service-user/src/main/java/com/bank/serviceuser/exception/EmailAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.bank.serviceuser.exception;
-
-public class EmailAlreadyExistsException extends RuntimeException {
-    public EmailAlreadyExistsException(String message) {
-        super(message);
-    }
-}
