--- service-user/src/main/java/com/bank/serviceuser/exception/FieldValidationError.java
@@ -1,19 +1,0 @@
-package com.bank.serviceuser.exception;
-
-public class FieldValidationError {
-    private String field;
-    private String message;
-
-    public FieldValidationError(String field, String message) {
-        this.field = field;
-        this.message = message;
-    }
-
-    public String getField() {
-        return field;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
