--- service-user/src/main/java/com/bank/serviceuser/exception/InvalidDniFormatException.java
@@ -1,7 +1,0 @@
-package com.bank.serviceuser.exception;
-
-public class InvalidDniFormatException extends RuntimeException {
-    public InvalidDniFormatException(String message) {
-        super(message);
-    }
-}
