--- service-user/src/main/java/com/bank/serviceuser/exception/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package com.bank.serviceuser.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
