--- service-user/src/main/java/com/bank/serviceuser/repository/CredentialRepository.java
@@ -1,9 +1,0 @@
-package com.bank.serviceuser.repository;
-
-import com.bank.serviceuser.model.Credencial;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface CredentialRepository extends JpaRepository<Credencial, Long> {
-    Optional<Credencial> findByIdUsuario(Long idUsuario);
-}
