--- service-user/src/main/java/com/bank/serviceuser/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.bank.serviceuser.repository;
-
-import com.bank.serviceuser.model.Usuario;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<Usuario, Long> {
-    Optional<Usuario> findByEmail(String email);
-    Optional<Usuario> findByDni(String dni);
-}
