--- service-user/src/main/java/com/bank/serviceuser/service/CredentialService.java
@@ -1,10 +1,0 @@
-package com.bank.serviceuser.service;
-
-import com.bank.serviceuser.model.Credencial;
-
-import java.util.Optional;
-
-public interface CredentialService {
-    Credencial createCredential(Credencial credencial);
-    Optional<Credencial> getCredentialByUserId(Long userId);
-}
