--- service-user/src/main/resources/application.properties
@@ -1,31 +1,0 @@
-# Puerto de ejecuci\u00f3n
-server.port=8081
-
-# Nombre del microservicio (para logs o Eureka si usas)
-spring.application.name=service-user
-
-# Configuraci\u00f3n de base de datos
-spring.datasource.url=jdbc:mysql://localhost:3306/banknet1
-spring.datasource.username=root
-spring.datasource.password=erfg545+DE
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-# JPA / Hibernate
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-
-# Zona horaria
-spring.jpa.properties.hibernate.jdbc.time_zone=UTC
-
-# Manejo de errores
-server.error.include-message=always
-server.error.include-binding-errors=always
-server.error.include-stacktrace=on_param
-
-# Codificaci\u00f3n UTF-8
-spring.http.encoding.charset=UTF-8
-spring.http.encoding.enabled=true
-spring.http.encoding.force=true
-
