--- service-user/src/test/java/com/bank/service_user/ServiceUserApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bank.service_user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ServiceUserApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
