+++ serviceuser/src/main/java/com/bank/serviceuser/ServiceuserApplication.java
@@ -1,0 +1,13 @@
+package com.bank.serviceuser;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ServiceuserApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ServiceuserApplication.class, args);
+	}
+
+}
