+++ serviceuser/src/main/java/com/bank/serviceuser/dto/UserResponseDTO.java
@@ -1,0 +1,12 @@
+package com.bank.serviceuser.dto;
+
+import lombok.Data;
+
+@Data
+public class UserResponseDTO {
+    private Long id;
+    private String nombreCompleto;
+    private String email;
+    private String telefono;
+}
+
