+++ serviceuser/src/main/java/com/bank/serviceuser/exception/FieldValidationError.java
@@ -1,0 +1,24 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+
+package com.bank.serviceuser.exception;
+
+public class FieldValidationError {
+    private String field;
+    private String message;
+
+    public FieldValidationError(String field, String message) {
+        this.field = field;
+        this.message = message;
+    }
+
+    public String getField() {
+        return field;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
