+++ serviceuser/src/main/java/com/bank/serviceuser/exception/InvalidDniFormatException.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+
+package com.bank.serviceuser.exception;
+
+public class InvalidDniFormatException extends RuntimeException {
+    public InvalidDniFormatException(String message) {
+        super(message);
+    }
+}
+
