+++ serviceuser/src/main/java/com/bank/serviceuser/repository/CredentialRepository.java
@@ -1,0 +1,14 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+
+package com.bank.serviceuser.repository;
+
+import com.bank.serviceuser.model.Credencial;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface CredentialRepository extends JpaRepository<Credencial, Long> {
+    Optional<Credencial> findByIdUsuario(Long idUsuario);
+}
