+++ serviceuser/src/test/java/com/bank/serviceuser/ServiceuserApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bank.serviceuser;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServiceuserApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
